refactor(viewer): extracted viewer actions & reducers
